requirements.txt
kivy
yfinance
google-generativeai
pandas
